+++ src/app/components/study/factors/factor/factor.component.ts
@@ -24,6 +24,9 @@ export class FactorComponent implements OnInit {
 
 	@Output() addFactorToSampleSheet = new EventEmitter<any>();
 
+	@select(state => state.study.readonly) studyReadonly;
+	isStudyReadOnly: boolean = false;
+
 	validationsId = 'factors.factor';
 
 	isModalOpen: boolean = false;
@@ -39,7 +42,12 @@ export class FactorComponent implements OnInit {
 	constructor( private fb: FormBuilder, private editorService: EditorService, private ngRedux: NgRedux<IAppState>) {
 		this.studyValidations.subscribe(value => { 
 	      	this.validations = value;
-	    });
+		});
+		this.studyReadonly.subscribe(value => { 
+			if(value != null){
+				this.isStudyReadOnly = value
+			}
+		});
 	}
 
 	ngOnInit() {
@@ -66,14 +74,16 @@ export class FactorComponent implements OnInit {
 	}
 
 	openModal() {
-		if(this.addNewFactor){
-			this.factor = new MTBLSFactor()
-			if(this.factorTypeComponent){
-				this.factorTypeComponent.values = []
+		if(!this.isStudyReadOnly){
+			if(this.addNewFactor){
+				this.factor = new MTBLSFactor()
+				if(this.factorTypeComponent){
+					this.factorTypeComponent.values = []
+				}
 			}
+			this.initialiseForm()
+			this.isModalOpen = true
 		}
-		this.initialiseForm()
-		this.isModalOpen = true
 	}
 
 	initialiseForm() {
@@ -98,51 +108,57 @@ export class FactorComponent implements OnInit {
 	}
 
 	save() {
-		if(this.getFieldValue('factorName') != ''){
-			this.isFormBusy = true;
-			if(!this.addNewFactor){
-				this.editorService.updateFactor(this.factor.factorName, this.compileBody()).subscribe( res => {
-					this.updateFactors(res, 'Factor updated.')
-					this.addFactorToSampleSheet.next(this.factor);
-				}, err => {
-					this.isFormBusy = false
-				});
-			}else{
-				let tempFactor = this.compileBody()
-				this.editorService.saveFactor(tempFactor).subscribe( res => {
-					this.updateFactors(res, 'Factor saved.')
-					this.isModalOpen = false
-					this.addFactorToSampleSheet.next(tempFactor.factor);
-				}, err => {
-					this.isFormBusy = false
-				});
+		if(!this.isStudyReadOnly){
+			if(this.getFieldValue('factorName') != ''){
+				this.isFormBusy = true;
+				if(!this.addNewFactor){
+					this.editorService.updateFactor(this.factor.factorName, this.compileBody()).subscribe( res => {
+						this.updateFactors(res, 'Factor updated.')
+						this.addFactorToSampleSheet.next(this.factor);
+					}, err => {
+						this.isFormBusy = false
+					});
+				}else{
+					let tempFactor = this.compileBody()
+					this.editorService.saveFactor(tempFactor).subscribe( res => {
+						this.updateFactors(res, 'Factor saved.')
+						this.isModalOpen = false
+						this.addFactorToSampleSheet.next(tempFactor.factor);
+					}, err => {
+						this.isFormBusy = false
+					});
+				}
 			}
 		}
 	}
 
 	delete() {
-		this.editorService.deleteFactor(this.factor.factorName).subscribe( res => {
-				this.updateFactors(res, 'Factor deleted.')
-				this.isDeleteModalOpen = false
-				this.isModalOpen = false
-			}, err => {
-				this.isFormBusy = false
-		});
+		if(!this.isStudyReadOnly){
+			this.editorService.deleteFactor(this.factor.factorName).subscribe( res => {
+					this.updateFactors(res, 'Factor deleted.')
+					this.isDeleteModalOpen = false
+					this.isModalOpen = false
+				}, err => {
+					this.isFormBusy = false
+			});
+		}
 	}
 
 	updateFactors(data, message){
-		this.editorService.getFactors().subscribe( res => {
-			toastr.success( message, "Success", {
-				"timeOut": "2500",
-				"positionClass": "toast-top-center",
-				"preventDuplicates": true,
-				"extendedTimeOut": 0,
-				"tapToDismiss": false
-			});
-		})
-		this.form.markAsPristine()
-		this.initialiseForm();
-		this.isModalOpen = true;
+		if(!this.isStudyReadOnly){
+			this.editorService.getFactors().subscribe( res => {
+				toastr.success( message, "Success", {
+					"timeOut": "2500",
+					"positionClass": "toast-top-center",
+					"preventDuplicates": true,
+					"extendedTimeOut": 0,
+					"tapToDismiss": false
+				});
+			})
+			this.form.markAsPristine()
+			this.initialiseForm();
+			this.isModalOpen = true;
+		}
 	}
 
 	compileBody() {
